feat: creacion de test para node y php imagenes
