+++ conversion-service/node-converter/package.json
@@ -3,9 +3,9 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node src/index.js",
-    "dev": "nodemon --watch src --exec \"node src/index.js\""
+    "dev": "nodemon --watch src --exec \"node src/index.js\"",
+    "test": "jest --runInBand"
   },
   "keywords": [],
   "author": "",
@@ -17,6 +17,8 @@
     "uuid": "^11.1.0"
   },
   "devDependencies": {
-    "nodemon": "^3.1.10"
+    "jest": "^29.7.0",
+    "nodemon": "^3.1.10",
+    "supertest": "^7.1.1"
   }
 }
