+++ conversion-service/php-converter/composer.json
@@ -1,9 +1,15 @@
 {
     "name": "groot/php-converter",
+    "scripts": {
+        "test": "phpunit --configuration phpunit.xml"
+    },
     "require": {
         "slim/slim": "^4.14",
         "slim/psr7": "^1.7",
         "nyholm/psr7": "^1.8",
         "guzzlehttp/guzzle": "^7.9"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^11.5"
     }
 }
