+++ conversion-service/php-converter/phpunit.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<phpunit bootstrap="vendor/autoload.php">
+  <testsuites>
+    <testsuite name="Conversion Tests">
+      <directory>./tests</directory>
+    </testsuite>
+  </testsuites>
+</phpunit>
